in progress

+++ src/main/java/DAL/DAO.java
@@ -7,5 +7,5 @@ public interface DAO<T> {
     public void update(T objet) throws DALException;
     public void delete(T objet)throws DALException;
     public List<T> selectAll() throws DALException;
-    public T selectById(long id ) throws DALException;
+    public T selectById(int id ) throws DALException;
 }
